reformat impt functions to accept numpy arrays directly instead of file names. replaced main.py by combining multi_mlx_2 and i2c_rpi_to_mlx
